wip: split clients doctor
